hRUV - intra- and hierarchical inter-batch normalisation.

Hierarchical approach to remove unwanted variation across multiple batches of data.

For intra method, rlm or loess methods fits robust linear model or loess respectively to samples selected from pCtName. rlmShort and loessShort both performs robust smoother rlm and loess respectively then performs RUV-III utilising the replicates identified from intra_rep. Parameters negCtl and intra_k will be used for RUV-III.

hruv(
  dat_list,
  assay,
  intra = c("rlm", "loess", "rlmShort", "loessShort"),
  inter = c("balanced", "concatenate"),
  intra_k = 5,
  inter_k = intra_k,
  pCtlName,
  negCtl,
  intra_rep,
  inter_rep,
  hOrder = NULL,
  newAssay = NULL
)

Arguments

dat_list

A list of SummarizedExperiment data where each object represents a batch.

assay

An assay name to measure the missingness of the signals.

intra

An intra-batch normalisation method. See Description for details.

inter

An inter-batch normalisation method. Either "balanced" or "concatenate" method for hierarchical approach.

intra_k

An intra-batch RUV-III parameter k value.

inter_k

An inter-batch RUV-III parameter k value.

pCtlName

A name of the variable in the colData of each SummarizedExperiment data in the dat_list. This variable in colData should be a numeric ID of which samples to use for robust smoother fitting.

negCtl

A vector of row IDs for selection of stable matbolites to be used as negative control in RUV.

intra_rep

A name of the variable in the colData of each SummarizedExperiment data in the dat_list. This variable in colData should be a logical vector of which samples are an intra-batch replicate

inter_rep

A name of the variable in the colData of each SummarizedExperiment data in the dat_list. This variable in colData should be a logical vector of which samples are an inter-batch replicate

hOrder

A vector of batch names from names of list dat_list.

newAssay

A name of the new assay for cleaned (preprocessed) data.

Value

A normalised data as SummarizedExperiment object.